+++ src/main/java/com/chaosbuffalo/mkultra/core/abilities/CleansingSeed.java
@@ -35,7 +35,7 @@ public class CleansingSeed extends PlayerAbility {
     }
 
     @Override
-    public int getManaCost(int currentRank) {
+    public float getManaCost(int currentRank) {
         return 1 + 2 * currentRank;
     }
