+++ src/main/java/com/chaosbuffalo/mkultra/core/abilities/ExplosiveGrowth.java
@@ -45,7 +45,7 @@ public class ExplosiveGrowth extends PlayerAbility {
     }
 
     @Override
-    public int getManaCost(int currentRank) {
+    public float getManaCost(int currentRank) {
         return 6 + currentRank * 2;
     }
